image/requirements.txt
opencv_python==4.6.0.66
Pillow==9.3.0
pytesseract==0.3.10
